some refactoring with place chat and dependies
