+++ src/index.js
@@ -1,101 +1,8 @@
-import MyGPT from "./ClientGPT.js"
-import { downloadFile, transformeImageToBase64, writeFile } from "./helpers.js";
-import MessageHistory from "./MessageHistory.js"
+import { Chat } from "./ChatClient.js";
+import { transformeImageToBase64, getterFile } from "./helpers.js";
 import TelegramBot from "node-telegram-bot-api"
 
 const ADMIN_ID = process.env.ADMIN_TELEGRAM_ID
-
-const placeHOlder = [
-  {
-    file_id: 'AgACAgIAAxkBAAIBoGg8vaWbJGsbYW_3qfjsLlmq1g8BAAJt9DEbcdLgSWoEdnvk0EwcAQADAgADeQADNgQ',
-    file_unique_id: 'AQADbfQxG3HS4El-',
-    file_size: 69447,
-    width: 1026,
-    height: 248
-  },
-  {
-    file_id: 'AgACAgIAAxkBAAIBs2g8vnv7lJRwvm82502-NMPUkONaAAJ69DEbcdLgSRkgIL6KA-H7AQADAgADeAADNgQ',
-    file_unique_id: 'AQADevQxG3HS4El9',
-    file_size: 8301,
-    width: 800,
-    height: 114
-  },
-  {
-    file_id: 'AgACAgIAAxkBAAIBpGg8vbLHETzkQtYHBKCmhSKEDmp9AAJu9DEbcdLgSR4QNIY6-IloAQADAgADeAADNgQ',
-    file_unique_id: 'AQADbvQxG3HS4El9',
-    file_size: 26806,
-    width: 658,
-    height: 201
-  },
-  {
-    file_id: 'AgACAgIAAxkBAAIBpWg8vcgTDtfOQMo06svmX9-sJs3KAAJv9DEbcdLgSWDiE537C5WdAQADAgADeQADNgQ',
-    file_unique_id: 'AQADb_QxG3HS4El-',
-    file_size: 47352,
-    width: 1132,
-    height: 583
-  },
-  {
-    file_id: 'AgACAgIAAxkBAAIBpmg8vciHoC16oiaTwqSgvpGkBULPAAJw9DEbcdLgSYxuX0NxrmHUAQADAgADeAADNgQ',
-    file_unique_id: 'AQADcPQxG3HS4El9',
-    file_size: 23715,
-    width: 519,
-    height: 515
-  },
-  {
-    file_id: 'AgACAgIAAxkBAAIB8Wg8xnjtAl9EiKaKfdO626tLgQMlAAKn9DEbcdLgSbfsEcv2MA-IAQADAgADeQADNgQ',
-    file_unique_id: 'AQADp_QxG3HS4El-',
-    file_size: 37485,
-    width: 928,
-    height: 502
-  }
-]
-const PHOTO_ID = "AgACAgIAAxkBAAIBnWg8rRhKdc3ASwy1383aFwjGS_IkAAIR9DEbcdLgSeQaEnDuRhIdAQADAgADeAADNgQ"
-
-class Chat {
-	chatID = null;
-	historyMessages = null;
-	myGPT = null;
-	gettingFiles = false;
-	allFiles = [];
-	messageFiles = [];
-	hotFiles = null;
-
-	constructor({chatID, saving = true}){
-		this.chatID = chatID;
-		this.historyMessages = new MessageHistory({save: saving});
-		this.myGPT = new MyGPT({
-			apiKey: process.env.OPEN_AI_API_KEY, 
-			history: this.historyMessages
-		});
-	}
-
-	// working with sending photos
-	addFile = function (obj) {
-		this.allFiles.push(obj)
-		this.messageFiles.push(obj)
-	};
-	clearFiles = function () {
-		this.messageFiles = []
-	};
-	getStatusFiles = function (){
-		return this.gettingFiles;
-	};
-	startFileSession = function () {
-		this.gettingFiles = true
-	};
-	endFilesSession = function () {
-		this.gettingFiles = false
-	}
-	// end working with sending photos
-}
-
-
-async function getterFile (data) {
-	const dataImage = await data
-	const format = dataImage.file_path.match(/(?<=\.)\w*$/)?.[0]
-	// // place for error
-	return downloadFile(`http://api.telegram.org/file/bot${process.env.TELEGRAM_TOKEN}/${dataImage.file_path}`, `${dataImage.file_unique_id}.${format}`)
-}
 
 async function workerCommand ({msg, chats, bot}){
 	const chatID = msg.chat.id;
@@ -106,7 +13,6 @@ async function workerCommand ({msg, chats, bot}){
 	if(!messageText) return 
 
 	if(messageText === '/start'){
-		const historyMessages = 
 		chats[chatID] = new Chat ({chatID})
 		return bot.sendMessage(chatID, "Чат успішно створений")
 	} 
@@ -118,7 +24,6 @@ async function workerCommand ({msg, chats, bot}){
 	
 	switch (messageText) {
 		case "/history":
-			
 			bot.sendMessage(chatID, "Готується істрія ції сессії")
 			bot.sendMessage(chatID, JSON.stringify(chat.historyMessages.getHistory(), null, 4))
 			bot.sendMessage(chatID, "Це вся історія цієї сессії") 
@@ -138,7 +43,7 @@ async function workerCommand ({msg, chats, bot}){
 			bot.sendMessage(chatID, "Файли готові, можете писати промпт")
 			break;
 		default:
-			bot.sendMessage(chatID, "Команда")
+			bot.sendMessage(chatID, "Цієї команди не знаю")
 			break;
 	}
 }
@@ -161,8 +66,6 @@ async function workerTextGPT ({msg, chats, bot}){
 }
 
 
-
-
 const initTelegram = () => {
 	const chats = {}
 
@@ -178,7 +81,6 @@ const initTelegram = () => {
 		const chat = chats[msg.chat.id]
 		if(!chat) return 
 
-
 		if(chat.getStatusFiles()) {
 			console.log('files', chat.getStatusFiles())
 			if(msg.photo){
@@ -188,14 +90,8 @@ const initTelegram = () => {
 		}
 
 		if(msg.text) return workerTextGPT({msg, chats, bot})
-
-		if(msg.media_group_id) {
-
-		}
 	})
-
-
+	return "some"
 }
-
 
 initTelegram()
